+++ elements/map/stories/getFeature-tooltip.js
@@ -44,15 +44,16 @@ const GetFeatureInfoTooltipStory = {
     ],
     center: [-10997148, 4569099],
     zoom: 4,
+    style: "width: 100%; height: 300px;",
+    storySlotContent: `<eox-map-tooltip></eox-map-tooltip>`,
   },
   render: /** @param {Object.<string, unknown>} args **/ (args) => html`
     <eox-map
-      id="tooltipTest"
-      style="width: 100%; height: 300px;"
       .center=${args.center}
       .controls=${args.controls}
       .layers=${args.layers}
       .zoom=${args.zoom}
+      style=${args.style}
     >
       <eox-map-tooltip></eox-map-tooltip>
     </eox-map>