+++ elements/map/stories/wmts-capabilities-layer.js
@@ -22,6 +22,7 @@ const WMTSCapabilitiesLayerStory = {
       },
     ],
     zoom: 5,
+    storyCodeBefore: `import "@eox/map/src/plugins/advancedLayersAndSources"`,
   },
 };